build: add pmproxy to the set needing LOCAL_ONLY port setup
Updates the RPM specs to match long-standing requirements not
to start daemons listening to open ports (pmproxy was missing
from the handling done for pmcd and pmlogger).  Also sync up
the pcp.spec.in and redhat.spec files in the way they handle
this requirement - picked "best" of both and this is managed
consistently now.