build: update pcp container to ensure needed ports are open
Since switching to a Fedora container as the base distro for
the PCP container, there's been inconsistent handling of the
ports for pmcd and pmproxy (with pmcd LOCAL_ONLY=1 now set if
building via Makepkgs).

Minor cleanup in terms of specifying both ports in one EXPOSE
line in the container spec, and accidental whitespace removal.