package-list and Debian builds: need systemd-dev
Our configure uses pkg-config to get systemd paths and if this
fails ENABLE_SYSTEMD is no ... which is bad.

The pkg-config .pc file for systemd is hidden in the systemd-dev
package (for Debian-based platforms), but is in the systemd
package for (RH-at-least) rpm-based platforms.

There is another commit that fixes the debian control file so
systemd-dev ends up in the Build-Depends clause.