source code changes for OpenIndiana
Reactivating the OpenIndiana platfrom in the QA Farm exposed some
bit-rot.

Our "exports" handing for libraries (.so.N files) was kinda broken:
- did not work at all for OpenIndiana because of different "comment"
  syntax
- we had symbols in exports files that did not appear in the
  library (usually old ones we've retired or interim ones from
  development that had been accidentally left behind) ... this
  are silently ignored for Linux, but are fatal for OpenIndiana
- the #if mechanism we added for generating exports from exports.in
  was not strong enough ... we need to be able to mimic the
  conditionals in the GNUmakefile where whole source files are
  included in the build or not; now
    .#if ENABLE_FOO
  is supported for the the ENABLE_FOO thingies in builddefs
These changes touch:
    src/libpcp/src/.gitignore
    src/libpcp/src/exports.in
    src/libpcp/src/mk.exports
    src/libpcp3/src/exports.in
    src/libpcp3/src/mk.exports
    src/libpcp_web/src/exports

Stop using index() and rindex() (these are from *BSD and removed from
POSIX.1-2008) ... strchr() and strrchr() will work just as well and
this is not only more standards-compliant, but also avoids the messy
    .#ifdef HAVE_STRINGS_H
    .#include <strings.h>
    .#endif
which in some cases was missing the required conditional guards.
This change touches:
    qa/src/churnctx.c
    qa/src/pdu-server.c
    src/include/pcp/platform_defs.h.in
    src/libpcp/src/derive_parser.y.in
    src/libpcp/src/exec.c
    src/libpcp3/src/derive_parser.y.in
    src/libpcp3/src/exec.c
    src/libpcp3/src/include/pcp/platform_defs.h.in
    src/libpcp3/src/spec.c
    src/pmdas/linux_proc/proc_pid.c
    src/pmdas/mmv/mmvdump.c
    src/pmdas/sample/src/proc.c
    src/pmie/src/pmie_dump_stats.c
    src/pmie/src/pragmatics.c
    src/pmlogmv/pmlogmv.c

OpenIndiana needs -D_POSIX_PTHREAD_SEMANTICS not just for the pthreads
stuff, but to get posix-compliant prototypes for things like ctime_r().
These changes touch:
    src/include/builddefs.in
    src/libpcp/src/GNUlibrarydefs
    src/libpcp3/src/GNUlibrarydefs
    src/libpcp_fault/src/GNUmakefile
    src/libpcp_static/src/GNUmakefile

I had introduced "local" in some shell scripts ... this is NOT portable,
so remove 'em, but it does require some shell variable renaming to avoid
problems in nested procedures.
This change touches:
    src/pmlogctl/pmlogctl.sh

There is a pmconfig in /usr/bin on OpenIndiana (the old Solaris
configure the Power Management system tool) which has nothing to
do with our pmconfig that's installed in /usr/libexec/pcp/bin here.
Force an absolute path to make sure we run the PCP one.
This change touches:
    src/pmlogger/pmlogger_daily.sh
    src/pmlogger/pmlogger_merge.sh
    src/pmlogger/pmlogger_rewrite.sh
    src/pmlogmv/pmlogmv.sh

And some miscellaneous ones ...
    qa/src/chk_memleak.c: logic error
    src/libpcp_web/src/loggroup.c: add missing #include <limits.h>
    src/pmcd/src/client.c: pander to gcc
    src/pmdas/mmv/src/mmv.c: pander to gcc