src/libpcp_web/src/exports: put back ini_parse
This symbol is conditional in the build, and I yanked it because
it did not appear on one platform I was looking at.

Putting it back now, pro tem.

Real fix is to use mk.exports as per libpcp, but that will need
extending (again) because the conditional guard in the makefile uses
a HAVE_FOO macro not an ENABLES_FOO macro.