libpcp_web: Eliminate a possible use after free in series_calculate_rate()
If the first argument of infofmt() is non-NULL, it attempts to use the
memory pointed at by the first argument.  However, there is a free()
in series_calculate_rate() that left msg as a non-NULL value after the
free(), causing a possible use after free.  The code has been
corrected so that msg variable is only non-NULL if it is pointing to
currently allocated memory.

Fixes CID 487579