src/pmlogrewrite/gram.y: rework use of assert()
assert() was being used here in ways that were not helpful, e.g.

    int	sts;
    sts = sscanf($1, "%d.%d.%d", &domain, &cluster, &item);
    assert(sts == 3);

Problems:

1. when NDEBUG is #defined, the assert() goes away and we're left
   with no test and a compiler warning for the unused variable sts.

2. the lexical scanner is already doing a lot of the checking,
   so in the case above, $1 is TOK_PMID_INT which must match the
   lexical pattern [0-9]+\.[0-9]+\.[0-9]+ so we can be sure sscanf
   will return 3 and domain >= 0 and cluster >=0 and item >=0

So the test can go away, and the following code checking for legal
values of domain, cluster and item can be simplified.

This pattern is repeated several times in the grammar, and this commit
cleans 'em all up.