pmcd: transition to simpler libbpf project usdt.h and macros
Updates the pmcd static probes for modern usage (bpftrace) and
significantly simplify the PCP bits in the process - some newer
PDUs were not reflected, now they are automatically since this
approach formats with libpcp routines.  No additional toolchain
needed anymore (dtrace) to compile, fewer package dependencies.

A sample bpftrace script is provided but as it is just a sample
and a developer is likely to be workin in-tree, do not ship it.