pmcd: transition to simpler libbpf project usdt.h and macros
Updates the pmcd static probes for modern usage (bpftrace) and
significantly simplify the PCP bits in the process - some newer
PDUs were not reflected, now they are automatically since this
approach formats with libpcp routines (conditionally, only when
probes are attached).  No extra build toolchain needs exist now
and as a result we have fewer package dependencies.

A sample bpftrace script is provided but as its just a sample,
and a developer is likely to be workin in-tree, and this newer
code require far less scaffolding than before, do not ship it.