libpcp: remove local strchrnul definition entirely
This is problematic on macOS platforms, as it is there for some
versions but not others, and we only build the one DMG image.
Conflicts exist within pcp-htop where htop tackles the problem
in a similar way (creating a local strchrnul) which complicates
the problem further if libpcp also defines it.

Fortunately, there are almost no uses and we can easily refactor
the two discover.c cases.