pmdaopenmetrics: fix benchmark performance bug
Resolves RHEL-124769. Clusters were being refreshed hundreds
of times / second because the configuration directory was
populated with sources that were not connected to the PMDA. This
caused the length of the metric dictionary to always equal 0 and
the conditional in the function old_enough_for_refresh to always
return true. The solution now only calls for refresh every
timeout seconds.