qa: Debian packing lists, redo perl-xs-dev
Turns out that up until trixie (and still today in sid) perl-xs-dev is
not a real package, and attempts to install it lead to libperl-dev
being installed; but this seems to be sufficient to satisfy the
(recent) conditional addition of perl-xs-dev in the Build-Depends:
clause for pcp.

Leaving perl-xs-dev in the packing lists means list-packages -m will
claim it is missing ... replacing it by libperl-dev in our packing
lists makes that problem go away and does not impact the builds
(AFAICT).