Debian build: rework post* and pre* scripts
As reported in
    https://bugs.launchpad.net/ubuntu/+source/pcp/+bug/2064784
the installation of PCP in a container may fail because neither init(1)
nor systemd(1) is really running, although the related commands
update-rc.d (and invoke-rc.d) and/or systemctl may be present.

This commit reworks the guards so that we don't try to run commands
that will fail, e.g. systemctl in this scenario reports:
    System has not been booted with systemd as init system (PID 1).
    Can't operate.
    Failed to connect to bus: Host is down
and since these scripts (once packaged and installed) are run with
-e from the #! line causing them to exit on failure and dpkg fails
to install the packages.