Ubuntu builds: rework migration from misnamed *t64* packages
This for an upgrade from Ubunut packages to the PCP developer packages.

No combination of Breaks|Replaces seems to work, so I'm opting for the
"transitional package" approach as described here:
    https://wiki.debian.org/RenamingPackages

debian/fixcontrol has been reworked to only include this fluff if
we're building on a Ubuntu platform, so it shoould not impact other
Debian-based platforms.