Move service start|stop|status|... scripts
As part of the Debian build and packaging refactoring, I discovered
that our real "rc" scripts for (mostly) starting and stopping pmcd,
pmlogger, pmproxy and pmie were in a place that violates at least
the Debian rules.

Specifically, outside /etc/init.d most files in /etc are deemed to be
"conffiles", and their mode is forced to 0644.  This does not work well
for /etc/pcp/pm<foo>/rc which we're expecting to be executable from
both the /etc/init.d/pm<foo> scripts and the systemd pm<foo>.service
configs.

So, $PCP_SERVICES_DIR is introduced in this commit.  This
configure-driven variable follows the logic for $PCP_BINADM_DIR,
and uses configure's $libexecdir as the path prefix, followed
by pcp/services.  This is a platform-specific path, but on Linux
becomes /usr/lib/pcp/services.

Since this directory is only for service scripts, the "rc" part of the
name (which dates to Unix systems older than System V) can be dropped,
so /etc/pcp/pmcd/rc becomes /usr/lib/pcp/services/pmcd.

These scripts and their locations are all "hidden" in the
sense that users execute the wrappers, systemctl ... pm<foo> or
/etc/init.d/pm<foo> ...  The one exception is /etc/pcp/pmcd/rc.local
which is mentioned in the PCPIntro(1) man page, and a symbolic
link is now installed to maintain compatibility for this one.