libpcp_pmda: pmdaCacheResize() was busted
Calling pmdaCacheResize() duly set maxinst for the indom cache but
then this value was ignored when the internal instance identifier was
being generated, e.g. when calling pmdaCacheStoreKey(), so the internal
instance identifier could have been as large as 2^31-1 (the default).

Fix this.
Extend qa/src/keycache.c with a -r option to trigger a call to
pmdaCacheResize().
Extend qa/489 with a src/keycache -r 32768 test case.