configure.ac: pander to autoheader in Debian build
The Debian build now runs autoheader, but we do not use autoheader.

The problem is that autoheader takes the *first* AC_CONFIG_HEADERS()
in configure.ac and *rewrites* that header ... in our case this
corrupted src/include/pcp/config.h.in which flowed through to
src/include/pcp/config.h and an obscure qa/377 failure.

We now have
    AC_CONFIG_HEADERS(src/include/pcp/autoheader-dummmy.h)
early-on in configure.ac so autoheader can create this and no one
on the PCP side of the build fence gives a rat's arse.