debian/*.dirs: remove all the empty ones
We really don't need these files, as we have no "special" case
directories and all PCP-related directories are installed and
associated with the correct packages as a consequence of descendent
files named in the .install files (or the remainder in the case of
the pcp package).

For example:

$ dpkg-query -S /usr/lib/pcp/pmns /var/lib/pcp/testsuite/src
pcp: /usr/lib/pcp/pmns
pcp-testsuite: /var/lib/pcp/testsuite/src

And empty .dirs files are serving no useful purpose.

But over in the rpm build, we need a little care if the empty .dirs
files are removed.