libpcp: add flags field to __pmFILE
For archive repairing in pmlogcheck (the next commit) I need to
know if an archive has been decompressed when __pmFopen() is called
(usually from pmNewContext()).

The flags filed to __pmFILE holds a bit-field of state information,
currently only PM_FILE_PRE_DECOMPRESS or PM_FILE_DYNAMIC_DECOMPRESS
(both new) respectively for the cases where we run an external
decompression program to a tempory file, or where decompression
on-the-fly is available (currently only xv when the magic library
is there).

__pmFILE is private to libpcp.h and not exposed at the PMAPI, so
there are no compatibility issues here.