pmlogcheck: add archive repair facility
- better integrity checks
- wordsmithing for some error messages
- remake some old QA .out files as a result
- repair will truncate after last good record for .index, .meta
  and data volumes
- new QA for -r and -R options (interactive and non-interactive
  repair)

This addresses (my) most common problem of files being one block
short in the filesystem after a system crash, or QA VM being shutdown
without warning.

I used to have a script to do thos, but it was never released and
adding the functionality to pmlogcheck is cleaner and more robust.