shell scripts everywhere: rewrite "export FOO=bar"
We've tried very hard to keep Bash-ims out of our shell scripts
(and makefiles), but one has escaped and is sort of omnipresent.

Specifically this Bash-ism
    export FOO=bar
does not work on real Bourne shells or the older Unix sh(1).

There is a simple and portable translation:
    FOO=bar; export FOO

This was tripping up efforts to build PCP on Solaris 10.

Besides the changes to 360+ files (!), there is a (new)
scripts/fix-exports script that actually did most of the work,
and (new) qa/1646 that exercises scripts/fix-exports [not that
anyone else cares, I was just leveraging the QA infrastructure
to assist with script development and testing].