src/pcp/geolocate/pcp-geolocate.service.in: remove ConditionPathExists
This was just wrong.  This is a one shot service that needs to be
opt in and then run successfully after each systemd restart.

The problem with the ConditionPathExists clause is
that when the service is enabled it fails everytime after
$PCP_SYSCONF_DIR/labels/optional/geolocate exists, leaving the service
in a bad state and this turd in the journal:

    pcp-geolocate.service - PCP geographical location label
    service was skipped because of an unmet condition check
    (ConditionPathExists=!/etc/pcp/labels/optional/geolocate)

qa/9000 needed to be spoken to as a result.