src/pmlogger/pmlogger_janitor.sh: fix serious flaw
sagar.sagar AT oracle.com has reported a customer issue where
pmlogger_janitor is trying to kill a process that is not a pmlogger
(that's not good), although because we run all our pmlogger management
scripts as user $PCP_USER (not root) the kill failed to terminate
the innocent victim (that's good).

After some more triage I was able to reproduce in an extra test
in qa/1443.

The problem was that pmlogger_janitor was taking the
$PCP_TMP_DIR/pmlogger "map" files to be gospel.

Now pmlogger_janitor checks that the <pid> encoded in the name of the
map file is a running pmlogger process; if that's not the case then
pmlogger_janitor removes the map file, rather than signalling <pid>.