src/pmlogger/pmlogger_janitor.sh: ps -p filtering NQR for *BSD
More than NQR it was completely broken, leading to infant mortality
for any pmlogger that popped its head over the parapet.

In fixing that, discovered that the PMLOGGER*JANITOR variables from
$PCP_SYSCONFIG_DIR/pmlogger were not getting set correctly in
pmlogger_check.

And futher, the PMLOGGER*JANITOR variables should not even be in
the sibling $PCP_SYSCONFIG_DIR/pmlogger_farm file.