libpcp et al: expose "note" line from pmlogger port map files
The last line in the pmlogger port map file was added (a while ago)
to record the -m or -x annotation from the command line when pmlogger
is run.

The annotation values "pmlogger_check" and "reexec" are critical
to the correct operation of pmlogger_check, pmlogger_daily and
pmlogger_janitor.

So ...
- add a char *node field in the __pmLogPort structure in libpcp.h
  (not part of the PMAPI, so no compatibility issue)
- extend __pmLogFindLocalPorts() to fill in the node field when it
  scans a port map file
- export the values via a (new) pmcd.pmlogger.note metric
- fix up qa/src/multithread9.c to accommodate extra PMNS name below
  pmcd.pmlogger