Revert "src/perl/*/Makefile.PL: avoid bogus rpath for .so's"
This reverts commit 632bb7de5ddfefd6810743dfc949fb66376a0f7a.

This change was installing Perl .so files like
/usr/lib/x86_64-linux-gnu/perl5/5.30/auto/PCP/PMDA/PMDA.so
that had been stripped of the linking library information, e.g.
there is no libpcp.so nor libpcp_pmda.so:
$ $ ldd /usr/lib/x86_64-linux-gnu/perl5/5.30/auto/PCP/PMDA/PMDA.so
	linux-vdso.so.1 (0x00007ffe21db3000)
	libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007fea3ee7e000)
	/lib64/ld-linux-x86-64.so.2 (0x00007fea3f09f000)

This was observed for both dpkg and rpm builds.

We either need another fix, or a different hammer for the lintian
issues that triggered this.