qa/322 & 374: don't chmod $PCP_TMP_DIR/pmlogger on a *remote* system
These tests date from the precambrian time when the single-celled
$PCP_TMP_DIR/pmlogger directory had the sticky bit set (sometimes).

As a consequence, prior to this commit, these tests would:
    sudo chmod o+w $PCP_TMP_DIR/pmlogger
from a ssh script run as pcpqa@$host where $host was some random
other QA host.

Now if you're running QA at the same time on $host and have
check.callback in play, then an *arbitrary* test on $host risks
failing with this in the .out.bad

    /var/lib/pcp/tmp/pmlogger: wrong mode: expected 775 (from src/pmlogger/GNUmakefile), found 777

Jeez!

Now we launch the remote pmlogger using user $PCP_USER so we can
create the mapfile in $PCP_TMP_DIR/pmlogger without mucking with
the directory permissions.