src/pmdas/dm/dmmultipath.c: fix small memory leak
In dm_multipath_instance_refresh() info is a pointer to a local struct
and a calloc'd struct ... on some hosts in the QA Farm this led to
a mem leak that was tripped by qa/1188.

Fix is to free() the calloc'd one on all return paths if it has not
been stached by pmdaCache*().